Move growth rate constraints to an extension
