+++ temoa/components/limits.py
@@ -22,7 +22,6 @@ import temoa.components.geography as geography
 import temoa.components.technology as technology
 from temoa.components.utils import (
     Operator,
-    get_adjusted_existing_capacity,
     get_variable_efficiency,
     operator_expression,
 )
@@ -132,64 +131,6 @@ def limit_tech_output_split_average_constraint_indices(
         for r, p, t, o, op in model.output_split_annual_vintages
         if t not in model.tech_annual
         for v in model.output_split_annual_vintages[r, p, t, o, op]
-    }
-
-
-def limit_growth_capacity_indices(model: TemoaModel) -> set[tuple[Region, Period, Technology, str]]:
-    return {
-        (r, p, t, op)
-        for r, t, op in model.limit_growth_capacity.sparse_keys()
-        for p in model.time_optimize
-    }
-
-
-def limit_degrowth_capacity_indices(
-    model: TemoaModel,
-) -> set[tuple[Region, Period, Technology, str]]:
-    return {
-        (r, p, t, op)
-        for r, t, op in model.limit_degrowth_capacity.sparse_keys()
-        for p in model.time_optimize
-    }
-
-
-def limit_growth_new_capacity_indices(
-    model: TemoaModel,
-) -> set[tuple[Region, Period, Technology, str]]:
-    return {
-        (r, p, t, op)
-        for r, t, op in model.limit_growth_new_capacity.sparse_keys()
-        for p in model.time_optimize
-    }
-
-
-def limit_degrowth_new_capacity_indices(
-    model: TemoaModel,
-) -> set[tuple[Region, Period, Technology, str]]:
-    return {
-        (r, p, t, op)
-        for r, t, op in model.limit_degrowth_new_capacity.sparse_keys()
-        for p in model.time_optimize
-    }
-
-
-def limit_growth_new_capacity_delta_indices(
-    model: TemoaModel,
-) -> set[tuple[Region, Period, Technology, str]]:
-    return {
-        (r, p, t, op)
-        for r, t, op in model.limit_growth_new_capacity_delta.sparse_keys()
-        for p in model.time_optimize
-    }
-
-
-def limit_degrowth_new_capacity_delta_indices(
-    model: TemoaModel,
-) -> set[tuple[Region, Period, Technology, str]]:
-    return {
-        (r, p, t, op)
-        for r, t, op in model.limit_degrowth_new_capacity_delta.sparse_keys()
-        for p in model.time_optimize
     }
 
 
@@ -950,392 +891,6 @@ def limit_emission_constraint(
         sys.stderr.write(msg % (e, emission_limit))
         return Constraint.Skip
 
-    return expr
-
-
-def limit_growth_capacity_constraint_rule(
-    model: TemoaModel, r: Region, p: Period, t: Technology, op: str
-) -> ExprLike:
-    r"""Constrain ramp up rate of available capacity"""
-    return limit_growth_capacity(model, r, p, t, op, False)
-
-
-def limit_degrowth_capacity_constraint_rule(
-    model: TemoaModel, r: Region, p: Period, t: Technology, op: str
-) -> ExprLike:
-    r"""Constrain ramp down rate of available capacity"""
-    return limit_growth_capacity(model, r, p, t, op, True)
-
-
-def limit_growth_capacity(
-    model: TemoaModel, r: Region, p: Period, t: Technology, op: str, degrowth: bool = False
-) -> ExprLike:
-    r"""
-    Constrain the change of capacity available between periods.
-    Forces the model to ramp up and down the availability of new technologies
-    more smoothly. Has constant (seed, :math:`S_{r,t}`) and proportional
-    (rate, :math:`R_{r,t}`) terms. This can be defined for a technology group
-    instead of one technology, in which case, capacity available is summed over
-    all technologies in the group. In the first period, previous available
-    capacity :math:`\mathbf{CAPAVL}_{r,p,t}` is replaced by previous existing
-    capacity, if any can be found.
-
-    .. math::
-        :label: Limit (De)Growth Capacity
-
-            \begin{aligned}\text{Growth:}\\
-            &\mathbf{CAPAVL}_{r,p,t}
-            \quad \le, \ge, \text{or} = \quad
-            S_{r,t} + (1+R_{r,t}) \cdot \mathbf{CAPAVL}_{r,p_{prev},t}
-            \end{aligned}
-
-            \qquad \forall \{r, p, t\} \in \Theta_{\text{limit\_growth\_capacity}}
-
-
-            \begin{aligned}\text{Degrowth:}\\
-            &\mathbf{CAPAVL}_{r,p_{prev},t}
-            \quad \le, \ge, \text{or} = \quad S_{r,t} + (1+R_{r,t}) \cdot \mathbf{CAPAVL}_{r,p,t}
-            \end{aligned}
-
-            \qquad \forall \{r, p, t\} \in \Theta_{\text{limit\_degrowth\_capacity}}
-    """
-
-    regions = geography.gather_group_regions(model, r)
-    techs = technology.gather_group_techs(model, t)
-
-    growth = model.limit_degrowth_capacity if degrowth else model.limit_growth_capacity
-    rate = 1 + value(growth[r, t, op][0])
-    seed = value(growth[r, t, op][1])
-    cap_rpt = model.v_capacity_available_by_period_and_tech
-
-    # relevant r, p, t indices
-    cap_indices = {(_r, _p, _t) for _r, _p, _t in cap_rpt.keys() if _t in techs and _r in regions}
-    # periods the technology can have capacity in this region (sorted)
-    periods = sorted({_p for _r, _p, _t in cap_rpt})
-
-    if len(periods) == 0:
-        if p == model.time_optimize.first():
-            msg = (
-                'Tried to set {}rowthCapacity constraint {} but there are no periods where this '
-                'technology is available in this region. Constraint skipped.'
-            ).format('Deg' if degrowth else 'G', (r, t))
-            logger.warning(msg)
-        return Constraint.Skip
-
-    # Only warn in p0 so we dont dump multiple warnings
-    if p == periods[0]:
-        if seed == 0:
-            msg = (
-                'No constant term (seed) provided for {}rowthCapacity constraint {}. '
-                'No capacity will be built in any period following one with zero capacity.'
-            ).format('Deg' if degrowth else 'G', (r, t))
-            logger.info(msg)
-        gaps = [
-            _p
-            for _p in model.time_optimize
-            if _p not in periods and min(periods) < _p < max(periods)
-        ]
-        if gaps:
-            msg = (
-                'Constructing {}rowthCapacity constraint {} and there are period gaps in which'
-                'capacity cannot exist in this region ({}). Capacity in these periods '
-                'will be treated as zero which may cause infeasibility or other problems.'
-            ).format('Deg' if degrowth else 'G', (r, t), gaps)
-            logger.warning(msg)
-
-    # sum available capacity in this period
-    capacity = quicksum(cap_rpt[_r, _p, _t] for _r, _p, _t in cap_indices if _p == p)
-
-    if p == model.time_optimize.first():
-        # First future period. Grab available capacity in last existing period
-        p_prev = model.time_exist.last()
-        capacity_prev = sum(
-            get_adjusted_existing_capacity(model, _r, _t, _v)
-            * value(model.process_life_frac[_r, p_prev, _t, _v])
-            for _r, _t, _v in model.existing_capacity.sparse_keys()
-            if _r in regions
-            and _t in techs
-            and _v + value(model.lifetime_process[_r, _t, _v]) > p_prev
-        )
-    else:
-        # Otherwise, grab previous future period
-        p_prev = model.time_optimize.prev(p)
-        capacity_prev = quicksum(cap_rpt[_r, _p, _t] for _r, _p, _t in cap_indices if _p == p_prev)
-
-    if degrowth:
-        expr = operator_expression(capacity_prev, Operator(op), seed + capacity * rate)
-    else:
-        expr = operator_expression(capacity, Operator(op), seed + capacity_prev * rate)
-
-    # Check if any variables are actually included before returning
-    if isinstance(expr, bool):
-        return Constraint.Skip
-    return expr
-
-
-def limit_growth_new_capacity_constraint_rule(
-    model: TemoaModel, r: Region, p: Period, t: Technology, op: str
-) -> ExprLike:
-    r"""Constrain ramp up rate of new capacity deployment"""
-    return limit_growth_new_capacity(model, r, p, t, op, False)
-
-
-def limit_degrowth_new_capacity_constraint_rule(
-    model: TemoaModel, r: Region, p: Period, t: Technology, op: str
-) -> ExprLike:
-    r"""Constrain ramp down rate of new capacity deployment"""
-    return limit_growth_new_capacity(model, r, p, t, op, True)
-
-
-def limit_growth_new_capacity(
-    model: TemoaModel, r: Region, p: Period, t: Technology, op: str, degrowth: bool = False
-) -> ExprLike:
-    r"""
-    Constrain the change of new capacity deployed between periods.
-    Forces the model to ramp up and down the deployment of new technologies
-    more smoothly. Has constant (seed, :math:`S_{r,t}`) and proportional
-    (rate, :math:`R_{r,t}`) terms. This can be defined for a technology group
-    instead of one technology, in which case, new capacity is summed over
-    all technologies in the group. In the first period, previous new capacity
-    :math:`\mathbf{NCAP}_{r,t,v_prev}` is replaced by previous existing capacity,
-    if any can be found.
-
-    .. math::
-        :label: Limit (De)Growth New Capacity
-
-            \begin{aligned}\text{Growth:}\\
-            &\mathbf{NCAP}_{r,t,v}
-            \quad \le, \ge, \text{or} = \quad
-            S_{r,t} + (1+R_{r,t}) \cdot \mathbf{NCAP}_{r,t,v_{prev}}
-            \text{ where } v=p
-            \end{aligned}
-
-            \qquad \forall \{r, p, t\} \in \Theta_{\text{limit\_growth\_capacity}}
-
-            \begin{aligned}\text{Degrowth:}\\
-            &\mathbf{NCAP}_{r,t,v_{prev}}
-            \quad \le, \ge, \text{or} = \quad S_{r,t} + (1+R_{r,t}) \cdot \mathbf{NCAP}_{r,t,v}
-            \text{ where } v=p
-            \end{aligned}
-
-            \qquad \forall \{r, p, t\} \in \Theta_{\text{limit\_degrowth\_capacity}}
-    """
-
-    regions = geography.gather_group_regions(model, r)
-    techs = technology.gather_group_techs(model, t)
-
-    growth = model.limit_degrowth_new_capacity if degrowth else model.limit_growth_new_capacity
-    rate = 1 + value(growth[r, t, op][0])
-    seed = value(growth[r, t, op][1])
-    new_cap_rtv = model.v_new_capacity
-
-    # relevant r, t, v indices
-    cap_rtv = {(_r, _t, _v) for _r, _t, _v in new_cap_rtv.keys() if _t in techs and _r in regions}
-    # periods the technology can be built in this region (sorted)
-    periods = sorted({_v for _r, _t, _v in cap_rtv})
-
-    if len(periods) == 0:
-        if p == model.time_optimize.first():
-            msg = (
-                'Tried to set {}rowthNewCapacity constraint {} but there are no periods where this '
-                'technology can be built in this region. Constraint skipped.'
-            ).format('Deg' if degrowth else 'G', (r, t))
-            logger.warning(msg)
-        return Constraint.Skip
-
-    # Only warn in p0 so we dont dump multiple warnings
-    if p == periods[0]:
-        if seed == 0:
-            msg = (
-                'No constant term (seed) provided for {}rowthNewCapacity constraint {}. '
-                'No capacity will be built in any period following one with zero new capacity.'
-            ).format('Deg' if degrowth else 'G', (r, t))
-            logger.info(msg)
-        gaps = [
-            _p
-            for _p in model.time_optimize
-            if _p not in periods and min(periods) < _p < max(periods)
-        ]
-        if gaps:
-            msg = (
-                'Constructing {}rowthNewCapacity constraint {} and there are period gaps in which'
-                'new capacity cannot be built in this region ({}). New capacity in these periods '
-                'will be treated as zero which may cause infeasibility or other problems.'
-            ).format('Deg' if degrowth else 'G', (r, t), gaps)
-            logger.warning(msg)
-
-    # sum new capacity in this period
-    new_cap = quicksum(new_cap_rtv[_r, _t, _v] for _r, _t, _v in cap_rtv if _v == p)
-
-    if p == model.time_optimize.first():
-        # First future period. Grab last existing vintage
-        p_prev = model.time_exist.last()
-        new_cap_prev = sum(
-            value(model.existing_capacity[_r, _t, _v])
-            for _r, _t, _v in model.existing_capacity.sparse_keys()
-            if _r in regions and _t in techs and _v == p_prev
-        )
-    else:
-        # Otherwise, grab previous future vintage
-        p_prev = model.time_optimize.prev(p)
-        new_cap_prev = sum(new_cap_rtv[_r, _t, _v] for _r, _t, _v in cap_rtv if _v == p_prev)
-
-    if degrowth:
-        expr = operator_expression(new_cap_prev, Operator(op), seed + new_cap * rate)
-    else:
-        expr = operator_expression(new_cap, Operator(op), seed + new_cap_prev * rate)
-
-    # Check if any variables are actually included before returning
-    if isinstance(expr, bool):
-        return Constraint.Skip
-    return expr
-
-
-def limit_growth_new_capacity_delta_constraint_rule(
-    model: TemoaModel, r: Region, p: Period, t: Technology, op: str
-) -> ExprLike:
-    r"""Constrain ramp up rate of change in new capacity deployment"""
-    return limit_growth_new_capacity_delta(model, r, p, t, op, False)
-
-
-def limit_degrowth_new_capacity_delta_constraint_rule(
-    model: TemoaModel, r: Region, p: Period, t: Technology, op: str
-) -> ExprLike:
-    r"""Constrain ramp down rate of change in new capacity deployment"""
-    return limit_growth_new_capacity_delta(model, r, p, t, op, True)
-
-
-def limit_growth_new_capacity_delta(
-    model: TemoaModel, r: Region, p: Period, t: Technology, op: str, degrowth: bool = False
-) -> ExprLike:
-    r"""
-    Constrain the acceleration of new capacity deployed between periods.
-    Forces the model to ramp up and down the change in deployment of new technologies
-    more smoothly. Has constant (seed, :math:`S_{r,t}`) and proportional
-    (rate, :math:`R_{r,t}`) terms. It is recommended to leave the rate term empty
-    as it would prevent the possibility of inflection in the rate of deployment.
-    This constraint can be defined for a technology group instead of one technology,
-    in which case, new capacity is summed over all technologies in the group. In the
-    first period, previous new capacities are replaced by previous existing capacities,
-    if any can be found.
-
-    .. math::
-        :label: Limit (De)Growth New Capacity Delta
-
-            \begin{aligned}\text{Growth:}\\
-            &\mathbf{NCAP}_{r,t,v_i} - \mathbf{NCAP}_{r,t,v_{i-1}}
-            \quad \le, \ge, \text{or} = \quad S_{r,t} + (1+R_{r,t}) \cdot
-            (\mathbf{NCAP}_{r,t,v_{i-1}} - \mathbf{NCAP}_{r,t,v_{i-2}})
-            \end{aligned}
-
-            \text{ where } v_i=p
-
-            \qquad \forall \{r, p, t\} \in \Theta_{\text{limit\_growth\_capacityDelta}}
-
-            \begin{aligned}\text{Degrowth:}\\
-            &\mathbf{NCAP}_{r,t,v_{i-1}} - \mathbf{NCAP}_{r,t,v_{i-2}}
-            \quad \le, \ge, \text{or} = \quad
-            S_{r,t} + (1+R_{r,t}) \cdot (\mathbf{NCAP}_{r,t,v_i} - \mathbf{NCAP}_{r,t,v_{i-1}})
-            \end{aligned}
-
-            \text{ where } v_i=p
-
-            \qquad \forall \{r, p, t\} \in \Theta_{\text{limit\_degrowth\_capacityDelta}}
-    """
-
-    regions = geography.gather_group_regions(model, r)
-    techs = technology.gather_group_techs(model, t)
-
-    growth = (
-        model.limit_degrowth_new_capacity_delta
-        if degrowth
-        else model.limit_growth_new_capacity_delta
-    )
-    rate = 1 + value(growth[r, t, op][0])
-    seed = value(growth[r, t, op][1])
-    new_cap_rtv = model.v_new_capacity
-
-    # relevant r, t, v indices
-    cap_rtv = {(_r, _t, _v) for _r, _t, _v in new_cap_rtv.keys() if _t in techs and _r in regions}
-    # periods the technology can be built in this region (sorted)
-    periods = sorted({_v for _r, _t, _v in cap_rtv})
-
-    if len(periods) == 0:
-        if p == model.time_optimize.first():
-            msg = (
-                'Tried to set {}rowthNewCapacityDelta constraint {} but there are no periods where '
-                'this technology can be built in this region. Constraint skipped.'
-            ).format('Deg' if degrowth else 'G', (r, t))
-            logger.warning(msg)
-        return Constraint.Skip
-
-    # Only warn in p0 so we dont dump multiple warnings
-    if p == periods[0]:
-        if seed == 0:
-            msg = (
-                'No constant term (seed) provided for {}rowthNewCapacityDelta constraint {}. '
-                'This is not recommended as deployment rates cannot inflect (change from '
-                'accelerating to decelerating or vice-versa).'
-            ).format('Deg' if degrowth else 'G', (r, t))
-            logger.warning(msg)
-        gaps = [
-            _p
-            for _p in model.time_optimize
-            if _p not in periods and min(periods) < _p < max(periods)
-        ]
-        if gaps:
-            msg = (
-                'Constructing {}rowthNewCapacityDelta constraint {} and there are period gaps in '
-                'which new capacity cannot be built in this region ({}). New capacity in these '
-                'periods will be treated as zero which may cause infeasibility or other problems.'
-            ).format('Deg' if degrowth else 'G', (r, t), gaps)
-            logger.warning(msg)
-
-    # sum new capacity in this period
-    new_cap = sum(new_cap_rtv[_r, _t, _v] for _r, _t, _v in cap_rtv if _v == p)
-
-    if p == model.time_optimize.first():
-        # First planning period, pull last two existing vintages
-        p_prev = model.time_exist.last()
-        new_cap_prev = sum(
-            value(model.existing_capacity[_r, _t, _v])
-            for _r, _t, _v in model.existing_capacity.sparse_keys()
-            if _r in regions and _t in techs and _v == p_prev
-        )
-        p_prev2 = model.time_exist.prev(p_prev)
-        new_cap_prev2 = sum(
-            value(model.existing_capacity[_r, _t, _v])
-            for _r, _t, _v in model.existing_capacity.sparse_keys()
-            if _r in regions and _t in techs and _v == p_prev2
-        )
-    else:
-        # Not the first future period. Grab previous future period
-        p_prev = model.time_optimize.prev(p)
-        new_cap_prev = sum(new_cap_rtv[_r, _t, _v] for _r, _t, _v in cap_rtv if _v == p_prev)
-        if p == model.time_optimize.at(2):  # apparently pyomo sets are indexed 1-based
-            # Second future period, grab last existing vintage
-            p_prev2 = model.time_exist.last()
-            new_cap_prev2 = sum(
-                value(model.existing_capacity[_r, _t, _v])
-                for _r, _t, _v in model.existing_capacity.sparse_keys()
-                if _r in regions and _t in techs and _v == p_prev2
-            )
-        else:
-            # At least the third future period. Grab last two future vintages
-            p_prev2 = model.time_optimize.prev(p_prev)
-            new_cap_prev2 = sum(new_cap_rtv[_r, _t, _v] for _r, _t, _v in cap_rtv if _v == p_prev2)
-
-    nc_delta_prev = new_cap_prev - new_cap_prev2
-    nc_delta = new_cap - new_cap_prev
-
-    if degrowth:
-        expr = operator_expression(nc_delta_prev, Operator(op), seed + nc_delta * rate)
-    else:
-        expr = operator_expression(nc_delta, Operator(op), seed + nc_delta_prev * rate)
-
-    # Check if any variables are actually included before returning
-    if isinstance(expr, bool):
-        return Constraint.Skip
     return expr
