+++ temoa/data_io/component_manifest.py
@@ -658,66 +658,6 @@ def build_manifest(model: TemoaModel, extension_ids: Sequence[str] | None = None
             is_table_required=False,
         ),
         LoadItem(
-            component=model.limit_growth_capacity,
-            table='limit_growth_capacity',
-            columns=['region', 'tech_or_group', 'operator', 'rate', 'seed'],
-            index_length=3,
-            validator_name='viable_rt',
-            validation_map=(0, 1),
-            is_period_filtered=False,
-            is_table_required=False,
-        ),
-        LoadItem(
-            component=model.limit_growth_new_capacity,
-            table='limit_growth_new_capacity',
-            columns=['region', 'tech_or_group', 'operator', 'rate', 'seed'],
-            index_length=3,
-            validator_name='viable_rt',
-            validation_map=(0, 1),
-            is_period_filtered=False,
-            is_table_required=False,
-        ),
-        LoadItem(
-            component=model.limit_growth_new_capacity_delta,
-            table='limit_growth_new_capacity_delta',
-            columns=['region', 'tech_or_group', 'operator', 'rate', 'seed'],
-            index_length=3,
-            validator_name='viable_rt',
-            validation_map=(0, 1),
-            is_period_filtered=False,
-            is_table_required=False,
-        ),
-        LoadItem(
-            component=model.limit_degrowth_capacity,
-            table='limit_degrowth_capacity',
-            columns=['region', 'tech_or_group', 'operator', 'rate', 'seed'],
-            index_length=3,
-            validator_name='viable_rt',
-            validation_map=(0, 1),
-            is_period_filtered=False,
-            is_table_required=False,
-        ),
-        LoadItem(
-            component=model.limit_degrowth_new_capacity,
-            table='limit_degrowth_new_capacity',
-            columns=['region', 'tech_or_group', 'operator', 'rate', 'seed'],
-            index_length=3,
-            validator_name='viable_rt',
-            validation_map=(0, 1),
-            is_period_filtered=False,
-            is_table_required=False,
-        ),
-        LoadItem(
-            component=model.limit_degrowth_new_capacity_delta,
-            table='limit_degrowth_new_capacity_delta',
-            columns=['region', 'tech_or_group', 'operator', 'rate', 'seed'],
-            index_length=3,
-            validator_name='viable_rt',
-            validation_map=(0, 1),
-            is_period_filtered=False,
-            is_table_required=False,
-        ),
-        LoadItem(
             component=model.limit_resource,
             table='limit_resource',
             columns=['region', 'tech_or_group', 'operator', 'cum_act'],
