+++ temoa/extensions/growth_rates/__init__.py
@@ -1,0 +1,3 @@
+from temoa.extensions.growth_rates.extension import GROWTH_RATES_EXTENSION
+
+__all__ = ['GROWTH_RATES_EXTENSION']
